requirements.txt
PyQt6==6.8.1
paddleocr==2.10.0
paddlepaddle==2.6.2
opencv-python==4.11.0.86
pillow==11.1.0
pyclipper==1.3.0.post6
fuzzywuzzy==0.18.0
keyboard==0.13.5
mss==10.0.0
PyAutoGUI==0.9.54
numpy==2.2.4
scipy==1.15.2
pandas==1.5.3
shapely==2.0.7
requests==2.32.3
pyinstaller==6.12.0